Found: fragment_changer.xml : Changed to more understandable parameter name.
SpinnerBindings.kt :
Changed setCurrencyValues to accept empty List and add logic to change the value if is null to emptyList.
Add Log.i to know the number of the list.

ChangerViewModel.kt :
Cleaned the code.
currencyList is getting the values direct from DB using LiveData.

ChangerRepository.kt
Add the method to return Currency list in a Livedata